pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spindl"
version = "0.1.0"
description = "Voice-activated AI agent with multimodal capabilities"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18.0",
    "pydantic>=2.0.0",
    "python-socketio>=5.10.0",
    "aiohttp>=3.9.0",
    "uvicorn>=0.27.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "psutil>=6.0.0",
    "chromadb>=0.5.0",
    "pillow>=10.0.0",
    "mss>=10.0.0",
    "pyvts>=0.3.3",
    "onnxruntime>=1.17.0",
    "tokenizers>=0.19.0",
    "twitchAPI>=4.2.0",
    "torch>=2.0.0",
    "sounddevice>=0.4.5",
    "kokoro>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
e2e = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.4",
    "python-socketio[asyncio_client]>=5.10.0",
    "aiohttp>=3.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "tests_e2e"]
addopts = "--junitxml=test-results.xml"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "vision: marks tests that require VLM capabilities",
    "cloud: marks tests that require cloud API credentials",
]

[tool.setuptools.packages.find]
where = ["src"]
